refactor: use native fieldset where applicable
- Rename FieldInput to Field
